feat: extract rs-dash-async crate from dash-sdk sync module
