+++ packages/rs-dash-async/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "dash-async"
+version.workspace = true
+edition = "2021"
+rust-version.workspace = true
+authors = ["Dash Core Team"]
+license = "MIT"
+description = "Async-sync bridging utilities for Dash Platform"
+
+[dependencies]
+thiserror = "2.0"
+tracing = "0.1.41"
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+tokio = { version = "1.40", features = ["rt", "rt-multi-thread"] }
+
+[dev-dependencies]
+tokio = { version = "1.40", features = ["macros", "rt-multi-thread"] }
