+++ packages/rs-sdk/Cargo.toml
@@ -18,6 +18,7 @@ drive = { path = "../rs-drive", default-features = false, features = [
 
 drive-proof-verifier = { path = "../rs-drive-proof-verifier", default-features = false }
 grovedb-commitment-tree = { git = "https://github.com/dashpay/grovedb", rev = "8f25b20d04bfc0e8bdfb3870676d647a0d74918b", features = ["client", "sqlite"], optional = true }
+dash-async = { path = "../rs-dash-async" }
 dash-context-provider = { path = "../rs-context-provider", default-features = false }
 dash-platform-macros = { path = "../rs-dash-platform-macros" }
 platform-encryption = { path = "../rs-platform-encryption" }
