+++ packages/rs-sdk/src/sync.rs
@@ -1,109 +1,17 @@
-//! Handle async calls from sync code.
-//!
-//! This is a workaround for an issue in tokio, where you cannot call `block_on` from sync call that is called
-//! inside a tokio runtime. This module spawns async futures in active tokio runtime, and retrieves the result
-//! using a channel.
+pub use dash_async::{block_on, AsyncError};
 
 use crate::error::Error;
-use dash_context_provider::ContextProviderError;
 use rs_dapi_client::{
     transport::sleep, update_address_ban_status, AddressList, CanRetry, ExecutionResult,
     RequestSettings,
 };
+use std::future::Future;
 use std::time::Duration;
-use std::{fmt::Debug, future::Future, sync::mpsc::SendError};
-
-#[derive(Debug, thiserror::Error)]
-pub enum AsyncError {
-    /// Not running inside tokio runtime
-    #[cfg(not(target_arch = "wasm32"))]
-    #[error("not running inside tokio runtime: {0}")]
-    NotInTokioRuntime(#[from] tokio::runtime::TryCurrentError),
-
-    /// Cannot receive response from async function
-    #[error("cannot receive response from async function: {0}")]
-    RecvError(#[from] std::sync::mpsc::RecvError),
-
-    /// Cannot send response from async function
-    #[error("cannot send response from async function: {0}")]
-    SendError(String),
-
-    #[error("asynchronous call from synchronous context failed: {0}")]
-    #[allow(unused)]
-    Generic(String),
-}
-
-impl<T> From<SendError<T>> for AsyncError {
-    fn from(error: SendError<T>) -> Self {
-        Self::SendError(error.to_string())
-    }
-}
-
-impl From<AsyncError> for ContextProviderError {
-    fn from(error: AsyncError) -> Self {
-        ContextProviderError::AsyncError(error.to_string())
-    }
-}
 
 impl From<AsyncError> for crate::Error {
     fn from(error: AsyncError) -> Self {
         Self::ContextProviderError(error.into())
     }
-}
-
-/// Blocks on the provided future and returns the result.
-///
-/// This function is used to call async functions from sync code.
-/// Requires the current thread to be running in a tokio runtime.
-///
-/// Due to limitations of tokio runtime, we cannot use `tokio::runtime::Runtime::block_on` if we are already inside a tokio runtime.
-/// This function is a workaround for that limitation.
-#[cfg(not(target_arch = "wasm32"))]
-pub fn block_on<F>(fut: F) -> Result<F::Output, AsyncError>
-where
-    F: Future + Send + 'static,
-    F::Output: Send,
-{
-    tracing::trace!("block_on: running async function from sync code");
-    let rt = tokio::runtime::Handle::try_current()?;
-    let (tx, rx) = std::sync::mpsc::channel();
-    tracing::trace!("block_on: Spawning worker");
-    let hdl = rt.spawn(worker(fut, tx));
-    tracing::trace!("block_on: Worker spawned");
-    let resp = tokio::task::block_in_place(|| rx.recv())?;
-
-    tracing::trace!("Response received");
-    if !hdl.is_finished() {
-        tracing::debug!("async-sync worker future is not finished, aborting; this should not happen, but it's fine");
-        hdl.abort(); // cleanup the worker future
-    }
-
-    Ok(resp)
-}
-
-#[cfg(target_arch = "wasm32")]
-pub fn block_on<F>(_fut: F) -> Result<F::Output, AsyncError>
-where
-    F: Future + Send + 'static,
-    F::Output: Send,
-{
-    unimplemented!("block_on is not supported in wasm");
-}
-
-/// Worker function that runs the provided future and sends the result back to the caller using mpsc channel.
-#[cfg(not(target_arch = "wasm32"))]
-async fn worker<F: Future>(
-    fut: F,
-    // response: oneshot::Sender<F::Output>,
-    response: std::sync::mpsc::Sender<F::Output>,
-) -> Result<(), AsyncError> {
-    tracing::trace!("Worker start");
-    let result = fut.await;
-    tracing::trace!("Worker async function completed, sending response");
-    response.send(result)?;
-    tracing::trace!("Worker response sent");
-
-    Ok(())
 }
 
 /// Retry the provided closure.
@@ -269,82 +177,6 @@ mod test {
             Arc,
         },
     };
-    use tokio::{
-        runtime::Builder,
-        sync::mpsc::{self, Receiver},
-    };
-
-    /// Test for block_on with async code that calls sync code, which then calls async code again.
-    ///
-    /// Given: An async function that calls a sync function, which then calls another async function.
-    /// When: The async function is executed using block_on.
-    /// Then: Other threads can still do some work
-    #[test]
-    fn test_block_on_nested_async_sync() {
-        let rt = Builder::new_multi_thread()
-            .worker_threads(1) // we should be good with 1 worker thread
-            .max_blocking_threads(1) // we should be good with 1 blocking thread
-            .enable_all()
-            .build()
-            .expect("Failed to create Tokio runtime");
-        // we repeat this test a few times, to make sure it's stable
-        for _repeat in 0..5 {
-            // Create a Tokio runtime; we use the current thread runtime for this test.
-
-            const MSGS: usize = 10;
-            let (tx, rx) = mpsc::channel::<usize>(1);
-
-            // Spawn new worker that will just count.
-            let worker = async move {
-                for count in 0..MSGS {
-                    tx.send(count).await.unwrap();
-                }
-            };
-            let worker_join = rt.spawn(worker);
-            // Define the number of levels of execution
-            let levels = 4;
-
-            // Define the innermost async function
-            async fn innermost_async_function(
-                mut rx: Receiver<usize>,
-            ) -> Result<String, ContextProviderError> {
-                for i in 0..MSGS {
-                    let count = rx.recv().await.unwrap();
-                    assert_eq!(count, i);
-                }
-
-                Ok(String::from("Success"))
-            }
-
-            // Define the nested sync function
-            fn nested_sync_function<F>(fut: F) -> Result<String, ContextProviderError>
-            where
-                F: Future<Output = Result<String, ContextProviderError>> + Send + 'static,
-                F::Output: Send,
-            {
-                block_on(fut)?.map_err(|e| ContextProviderError::Generic(e.to_string()))
-            }
-
-            // Define the outer async function
-            async fn outer_async_function(
-                levels: usize,
-                rx: Receiver<usize>,
-            ) -> Result<String, ContextProviderError> {
-                let mut result = innermost_async_function(rx).await;
-                for _ in 0..levels {
-                    result = nested_sync_function(async { result });
-                }
-                result
-            }
-
-            // Run the outer async function using block_on
-            let result = rt.block_on(outer_async_function(levels, rx));
-
-            rt.block_on(worker_join).unwrap();
-            // Assert the result
-            assert_eq!(result.unwrap(), "Success");
-        }
-    }
 
     use crate::error::StaleNodeError;
     use rs_dapi_client::DapiClientError;
@@ -407,12 +239,6 @@ mod test {
 
     /// Test that when we get "no available addresses" error, we return the last meaningful error
     /// wrapped in NoAvailableAddresses.
-    ///
-    /// This simulates the scenario where:
-    /// 1. First request fails with a meaningful error (e.g., stale node)
-    /// 2. The address gets banned
-    /// 3. On retry, there are no available addresses left
-    /// 4. We should return NoAvailableAddresses wrapping the original meaningful error
     #[tokio::test]
     async fn test_retry_returns_last_meaningful_error_on_no_addresses() {
         let call_count = Arc::new(AtomicUsize::new(0));
@@ -426,7 +252,6 @@ mod test {
             let count = call_count_clone.fetch_add(1, Ordering::Relaxed);
             async move {
                 if count == 0 {
-                    // First call: return a meaningful retryable error (stale node)
                     Err(ExecutionError {
                         inner: Error::StaleNode(StaleNodeError::Height {
                             expected_height: 100,
@@ -437,7 +262,6 @@ mod test {
                         address: Some("http://localhost:1".parse().unwrap()),
                     })
                 } else {
-                    // Subsequent calls: simulate no addresses available (all banned)
                     Err(ExecutionError {
                         inner: Error::DapiClientError(DapiClientError::NoAvailableAddresses),
                         retries: 0,
@@ -449,7 +273,6 @@ mod test {
 
         let result: ExecutionResult<(), Error> = retry(&address_list, settings, closure).await;
 
-        // We should get NoAvailableAddresses wrapping the stale node error
         let error = result.expect_err("should fail");
         match &error.inner {
             Error::NoAvailableAddressesToRetry(inner) => {
@@ -481,7 +304,6 @@ mod test {
         settings.retries = Some(5);
 
         let closure = move |_settings: RequestSettings| async move {
-            // First and only call returns "no available addresses"
             Err(ExecutionError {
                 inner: Error::DapiClientError(DapiClientError::NoAvailableAddresses),
                 retries: 0,
